+++ lindeck-frontend/package.json
@@ -27,6 +27,7 @@
     "react-markdown": "^5.0.2",
     "react-resize-detector": "^5.2.0",
     "react-scripts": "4.0.0",
+    "react-tabs": "^3.1.1",
     "uuid": "^8.3.1",
     "web-vitals": "^0.2.4"
   },
