+++ lindeck-frontend/src/views/boody/deckPage/card/cardEditDialog/CardEditDialog.css
@@ -1,6 +1,6 @@
 .inputCardText {
     background-color: #f8f8f8;
-    min-height: 300px;
+    min-height: 100px;
     width: 100%;
 }
